adding REST APIs

+++ backend/src/main/java/nemop/service/impl/JpaChannelService.java
@@ -9,12 +9,18 @@ import org.springframework.stereotype.Service;
 import nemop.model.Channel;
 import nemop.repository.ChannelRepository;
 import nemop.service.ChannelService;
+import nemop.support.ChannelDtoToChannel;
+import nemop.web.dto.ChannelDTO;
 
 @Service
 public class JpaChannelService implements ChannelService{
 
 	@Autowired
 	private ChannelRepository channelRepository;
+	
+	@Autowired
+	private ChannelDtoToChannel toChannel;
+	
 	
 	@Override
 	public Optional<Channel> findOne(Long id) {
@@ -27,14 +33,22 @@ public class JpaChannelService implements ChannelService{
 	}
 
 	@Override
-	public Channel save(Channel channel) {
-		return channelRepository.save(channel);
+	public Channel save(ChannelDTO channelDto) {
+		Channel channel = toChannel.convert(channelDto);
+		Channel savedChannel = channelRepository.save(channel);
+		return savedChannel;
 	}
 
 	@Override
-	public void delete(Long id) {
-		channelRepository.deleteById(id);
+	public Channel delete(Long id) {
+		Optional<Channel> channelOptional = channelRepository.findById(id);
+		if(channelOptional.isPresent()) {
+			Channel channel = channelOptional.get();
+			channelRepository.deleteById(id);
+			return channel;
+		}
 		
+		return null;
 	}
 
 	@Override
